+++ src/compy/export/BaseConfig.js
@@ -1,7 +1,7 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { MustBe } from '@this/typey';
+import { AskIf, MustBe } from '@this/typey';
 
 
 /**
@@ -105,6 +105,7 @@ export class BaseConfig {
         }
       } catch (e) {
         if (!hasConfig) {
+          // This could also be due to a bug in the config class.
           throw new Error(`Missing required configuration property: \`${name}\``);
         }
         throw e;
@@ -170,5 +171,79 @@ export class BaseConfig {
     }
 
     return result;
+  }
+
+
+  //
+  // Static members
+  //
+
+  /**
+   * "Evaluate" a configuration argument that was passed to the constructor of
+   * a configurable class. This is where the usual rules (e.g. as described by
+   * the `BaseComponent` constructor) are actually implemented. This method is
+   * expected to be called on a concrete subclass of this (base) class, and the
+   * actual called class is used to drive the salient portion of the error
+   * checking.
+   *
+   * @param {?object} rawConfig Raw configuration object, including allowing
+   *   `null` to be equivalent to `{}`, and accepting an instance of this class.
+   * @param {object} options Evaluation options.
+   * @param {object} [options.defaults] Default values when evaluating a plain
+   *   object. Defaults to `{}`.
+   * @param {function(new:*)} options.targetClass The class that `rawConfig` is
+   *   supposed to be constructing.
+   * @returns {BaseConfig} Instance of the concrete class that this method was
+   *   called on.
+   */
+  static eval(rawConfig, { defaults = {}, targetClass }) {
+    rawConfig ??= {};
+
+    if (rawConfig instanceof this) {
+      return rawConfig;
+    } else if (rawConfig instanceof BaseConfig) {
+      // It's the wrong concrete config class.
+      const gotName = rawConfig.constructor.name;
+      throw new Error(`Incompatible configuration class: expected ${this.name}, got ${gotName}`);
+    } else if (!AskIf.plainObject(rawConfig)) {
+      if (typeof rawConfig === 'object') {
+        const gotName = rawConfig.constructor.name;
+        throw new Error(`Cannot convert non-configuration object: expected ${this.name}, got ${gotName}`);
+      } else {
+        const gotType = typeof rawConfig;
+        throw new Error(`Cannot evaluate non-object as configuration: expected ${this.name}, got ${gotType}`);
+      }
+    }
+
+    // It's a plain object.
+
+    let configObj = rawConfig; // Might get replaced by a modified copy.
+
+    const defaultProp = (k, v, force = false) => {
+      if (force || !Reflect.has(configObj, k)) {
+        if (configObj === rawConfig) {
+          configObj = { ...rawConfig };
+        }
+        configObj[k] = v;
+      }
+    };
+
+    for (const [k, v] of Object.entries(defaults)) {
+      defaultProp(k, v);
+    }
+
+    const configTargetClass = configObj.class;
+
+    if ((configTargetClass === null) || (configTargetClass === undefined)) {
+      defaultProp('class', targetClass);
+    } else if (configTargetClass !== targetClass) {
+      if (!AskIf.constructorFunction(configTargetClass)) {
+        throw new Error('Expected class (constructor function) for `rawConfig.class`.');
+      } else {
+        throw new Error(`Mismatch on \`rawConfig.class\`: expected ${targetClass.name}, got ${configTargetClass.name}`);
+      }
+    }
+
+    return new this(configObj);
   }
 }