+++ src/compy/index.js
@@ -1,12 +1,14 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-export * from '#x/BaseAggregateComponent';
 export * from '#x/BaseComponent';
 export * from '#x/BaseConfig';
-export * from '#x/BaseThreadComponent';
+export * from '#x/BaseRootComponent';
 export * from '#x/BaseWrappedHierarchy';
 export * from '#x/ControlContext';
+export * from '#x/MockComponent';
+export * from '#x/MockRootComponent';
 export * from '#x/Names';
 export * from '#x/RootControlContext';
-export * from '#x/TemplRootComponent';
+export * from '#x/TemplAggregateComponent';
+export * from '#x/TemplThreadComponent';